Refactor ConnectionsMessageCollector
